+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/CourseService.java
@@ -28,35 +28,35 @@ public interface CourseService {
      * @param course modified course
      * @return true if the course was successfully updated, false otherwise
      */
-    boolean update(Integer id, Course course);
+    boolean update(Long id, Course course);
 
     /**
      * Attempts to delete a course
      * @param id of the course to be deleted
      * @return true if the course was successfully removed, false otherwise
      */
-    boolean delete(Integer id);
+    boolean delete(Long id);
 
     /**
      * Gets all the current available courses of a user
      * @param userId of the user to retrieve the courses from
      * @return list containing all the current available courses (if any)
      */
-    List<Course> list(Integer userId);
+    List<Course> list(Long userId);
 
     /**
      * Attempts to get a course given an id
      * @param courseId of the course to be retrieved
      * @return the course corresponding to the given id if it exists, null otherwise
      */
-    Optional<Course> getById(Integer courseId);
+    Optional<Course> getById(Long courseId);
 
     /**
      * Gets the map of teachers for the given course
      * @param courseId of the course to get the teachers from
      * @return map of Users as a key where the value is the teacher Role in that course
      */
-    Map<User, Role> getTeachers(Integer courseId);
+    Map<User, Role> getTeachers(Long courseId);
 
     /**
      * Returns if the user belongs to the course
@@ -64,7 +64,7 @@ public interface CourseService {
      * @param courseId of the course to check
      * @return true if the user belongs to the given course, false otherwise
      */
-    boolean belongs(Integer userId, Integer courseId);
+    boolean belongs(Long userId, Long courseId);
 
     /**
      * Returns true if the user is a teacher in the current course
@@ -72,7 +72,7 @@ public interface CourseService {
      * @param courseId of the course to check the privilege
      * @return true if the user is a teacher in the current course, false otherwise
      */
-    boolean isTeacher(Integer userId, Integer courseId);
+    boolean isTeacher(Long userId, Long courseId);
 
     /**
      * Returns true if the user is a helper in the current course
@@ -80,7 +80,7 @@ public interface CourseService {
      * @param courseId of the course to check the privilege
      * @return true if the user is a helper in the current course, false otherwise
      */
-    boolean isHelper(Integer userId, Integer courseId);
+    boolean isHelper(Long userId, Long courseId);
 
     /**
      * Returns true if the user is a student in the current course
@@ -88,7 +88,7 @@ public interface CourseService {
      * @param courseId of the course to check the privilege
      * @return true if the user is a student in the current course, false otherwise
      */
-    boolean isStudent(Integer userId, Integer courseId);
+    boolean isStudent(Long userId, Long courseId);
 
     /**
      * Returns true if the user is a helper or a teacher in the current course
@@ -96,5 +96,5 @@ public interface CourseService {
      * @param courseId of the course to check the privilege
      * @return true if the user is a helper or a teacher in the current course, false otherwise
      */
-    boolean isPrivileged(Integer userId, Integer courseId);
+    boolean isPrivileged(Long userId, Long courseId);
 }
