+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/TimetableDao.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface TimetableDao {
     boolean create(Course course, int dayOfWeek, Time start, Time end);
-    boolean update(int course_id, int dayOfWeek, Time start, Time end);
-    boolean delete(int course_id);
-    public List<Timetable> getById(long courseId);
+    boolean update(Long course_id, int dayOfWeek, Time start, Time end);
+    boolean delete(Long course_id);
+    public List<Timetable> getById(Long courseId);
 }
